requirements.txt
streamlit==1.24.0
openai==1.3.0
google-cloud-aiplatform==1.35.0
anthropic==0.3.0
python-dotenv==1.0.0 